backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyjwt==2.8.1
werkzeug==3.0.1
marshmallow==3.20.1
bleach==6.1.0
